server-parser/requirements.txt
django
djangorestframework
django-crontab
django-cron